+++ lib/src/clixon_yang.c
@@ -1504,7 +1504,7 @@ yang_parse_recurse(clicon_handle h,
 		   yang_spec    *ysp)
 {
     yang_stmt  *yi = NULL; /* import */
-    yang_stmt  *ymod;
+    yang_stmt  *ymod = NULL;
     yang_stmt  *yrev;
     char       *modname;
     char       *subrevision;
@@ -1527,7 +1527,7 @@ yang_parse_recurse(clicon_handle h,
 	    if ((nr = yang_parse_find_match(h, yang_dir, module, fbuf)) < 0)
 		goto done;
 	    if (nr == 0){
-		clicon_err(OE_YANG, errno, "No matching %s yang files found", module);
+		clicon_err(OE_YANG, errno, "No matching %s yang files found (expected modulenameor absolute filename)", module);
 		goto done;
 	    }
 	}